requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.0.14
django-ckeditor==6.1.0
django-heroku==0.3.1
django-js-asset==1.2.2
gunicorn==20.1.0
Pillow==8.3.0
psycopg2==2.8
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
